+++ d60.Cirqus.Tests/Contracts/EventStore/EventStoreTest.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics.Eventing.Reader;
 using System.Linq;
 using System.Text;
 using d60.Cirqus.Aggregates;